requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-wtf==1.2.1
wtforms==3.1.1
python-dotenv==1.0.0